refactor: rename mod
