+++ src/mint/mod.rs
@@ -1,2 +1,2 @@
-mod mint;
-pub use mint::{mint_loop, MintResult};
+mod miner;
+pub use miner::{mint_loop, MintResult};
